+++ api/src/main/java/team/pickz/api/domain/draft/application/DraftRoomService.java
@@ -37,8 +37,6 @@ public class DraftRoomService {
     private final MemberRepository memberRepository;
     private final DraftRoomRepository draftRoomRepository;
     private final DraftParticipantRepository draftParticipantRepository;
-    private final DraftRoomStreamerRepository draftRoomStreamerRepository;
-    private final RoomSequenceManager roomSequenceManager;
     private final ApplicationEventPublisher applicationEventPublisher;
 
     @Transactional
@@ -77,124 +75,6 @@ public class DraftRoomService {
                 .participantToken(host.getParticipantToken())
                 .isHost(true)
                 .build();
-    }
-
-    @Transactional
-    public void saveDraftRoomStreamers(Long roomId, String participantToken, List<DraftRoomStreamerRequest> requests) {
-        DraftParticipant requestor = draftParticipantRepository.findByRoomIdAndParticipantToken(roomId, participantToken)
-                .orElseThrow(() -> new IllegalArgumentException("유효하지 않은 참여자입니다."));
-
-        if (!requestor.isHost()) {
-            throw new IllegalArgumentException("방장만 스트리머 풀을 설정할 수 있습니다.");
-        }
-
-        draftRoomStreamerRepository.deleteAllByRoomId(roomId);
-
-        for (DraftRoomStreamerRequest req : requests) {
-            savePositionStreamers(roomId, req.teamSlot(), Position.TOP, req.top());
-            savePositionStreamers(roomId, req.teamSlot(), Position.JUG, req.jug());
-            savePositionStreamers(roomId, req.teamSlot(), Position.MID, req.mid());
-            savePositionStreamers(roomId, req.teamSlot(), Position.ADC, req.adc());
-            savePositionStreamers(roomId, req.teamSlot(), Position.SUP, req.sup());
-            savePositionStreamers(roomId, req.teamSlot(), Position.COACH, req.coach());
-        }
-    }
-
-    private void savePositionStreamers(Long roomId, int teamSlot, Position position, StreamerInfo info) {
-        if (info == null || info.name() == null || info.name().isBlank()) return;
-
-        DraftRoomStreamer streamer = DraftRoomStreamer.builder()
-                .roomId(roomId)
-                .teamSlot(teamSlot)
-                .position(position)
-                .streamerName(info.name())
-                .imageUrl(info.imageUrl())
-                .build();
-        draftRoomStreamerRepository.save(streamer);
-    }
-
-    @Transactional(readOnly = true)
-    public DraftRoomStreamerResponse getDraftRoomStreamers(Long roomId) {
-        List<DraftRoomStreamer> streamers = draftRoomStreamerRepository.findAllByRoomId(roomId);
-
-        List<StreamerInfo> top = extractByPosition(streamers, Position.TOP);
-        List<StreamerInfo> jungle = extractByPosition(streamers, Position.JUG);
-        List<StreamerInfo> mid = extractByPosition(streamers, Position.MID);
-        List<StreamerInfo> adc = extractByPosition(streamers, Position.ADC);
-        List<StreamerInfo> support = extractByPosition(streamers, Position.SUP);
-        List<StreamerInfo> coach = extractByPosition(streamers, Position.COACH);
-
-        return new DraftRoomStreamerResponse(top, jungle, mid, adc, support, coach);
-    }
-
-    private List<StreamerInfo> extractByPosition(List<DraftRoomStreamer> streamers, Position position) {
-        return streamers.stream()
-                .filter(s -> s.getPosition() == position)
-                .sorted(Comparator.comparingInt(DraftRoomStreamer::getTeamSlot))
-                .map(s -> new StreamerInfo(s.getStreamerName(), s.getImageUrl()))
-                .toList();
-    }
-
-    @Transactional
-    public ParticipantResponse joinRoom(String inviteCode) {
-        DraftRoom room = draftRoomRepository.findByInviteCode(inviteCode)
-                .orElseThrow(() -> new IllegalArgumentException("유효하지 않은 초대 코드입니다."));
-
-        if(room.getStatus() != RoomStatus.WAITING) {
-            throw new IllegalStateException("이미 게임이 시작된 방입니다.");
-        }
-
-        int sequence = roomSequenceManager.getNextSequence(room.getId());
-        String nickname = RandomNicknameGenerator.generate(sequence);
-
-        DraftParticipant participant = DraftParticipant.builder()
-                .roomId(room.getId())
-                .nickname(nickname)
-                .isHost(false)
-                .build();
-
-        draftParticipantRepository.save(participant);
-
-        List<DraftParticipant> participants = draftParticipantRepository.findAllByRoomId(room.getId());
-        List<String> nicknames = participants.stream()
-                .map(DraftParticipant::getNickname)
-                .toList();
-
-        ParticipantUpdateEvent event = ParticipantUpdateEvent.builder()
-                .totalCount(nicknames.size())
-                .nicknames(nicknames)
-                .newParticipant(nickname)
-                .build();
-
-        applicationEventPublisher.publishEvent(
-                ParticipantJoinedEvent.builder()
-                        .roomId(room.getId())
-                        .payload(event)
-                        .build()
-        );
-
-        return ParticipantResponse.builder()
-                .roomId(room.getId())
-                .participantToken(participant.getParticipantToken())
-                .isHost(participant.isHost())
-                .build();
-    }
-
-    @Transactional
-    public void selectCoach(Long roomId, String participantToken, String coachName, int targetTurnOrder) {
-        DraftParticipant participant = draftParticipantRepository.findByParticipantToken(participantToken)
-                .orElseThrow(() -> new IllegalArgumentException("유효하지 않은 참여자입니다."));
-
-        // 동시성 처리 고려: 이미 누군가 해당 감독/순서를 골랐는지 검증
-        boolean isAlreadySelected = draftParticipantRepository.existsByRoomIdAndSelectedCoachName(roomId, coachName);
-        if (isAlreadySelected) {
-            throw new IllegalStateException("이미 다른 참가자가 선택한 감독입니다.");
-        }
-
-        participant.selectCoach(coachName, targetTurnOrder);
-
-        // WebSocket 이벤트 발행 로직 (선택사항)
-        // applicationEventPublisher.publishEvent(...);
     }
 
     @Transactional
@@ -237,94 +117,5 @@ public class DraftRoomService {
                         .build()
         );
     }
-
-    @Transactional(readOnly = true)
-    public DraftPlayStateResponse getDraftPlayState(Long roomId) {
-        DraftRoom room = draftRoomRepository.findById(roomId)
-                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 방입니다."));
-
-        // 1. 참가자(감독) 정보 조회 및 정렬
-        List<DraftParticipant> participants = draftParticipantRepository.findAllByRoomId(roomId);
-
-        // 감독 선택을 완료한 참가자만 필터링 후 픽 순서(turnOrder)대로 정렬
-        List<DraftParticipant> coaches = participants.stream()
-                .filter(p -> p.getSelectedCoachName() != null)
-                .sorted(Comparator.comparing(p -> p.getTurnOrder() != null ? p.getTurnOrder() : 999))
-                .toList();
-
-        // "pickOrder": ["감독A", "감독B"...] 배열 추출
-        List<String> pickOrder = coaches.stream()
-                .map(DraftParticipant::getSelectedCoachName)
-                .toList();
-
-        // "coaches": [...] 객체 배열 추출
-        List<CoachResponse> coachResponses = coaches.stream()
-                .map(p -> new CoachResponse(
-                        p.getSelectedCoachName(),
-                        p.getNickname()
-                ))
-                .toList();
-
-        // 2. 스트리머 풀 조회 로직 재활용
-        List<DraftRoomStreamer> streamers = draftRoomStreamerRepository.findAllByRoomId(roomId);
-
-        DraftConfigResponse.StreamersByLine streamersByLine = new DraftConfigResponse.StreamersByLine(
-                extractByPosition(streamers, Position.TOP),
-                extractByPosition(streamers, Position.JUG),
-                extractByPosition(streamers, Position.MID),
-                extractByPosition(streamers, Position.ADC),
-                extractByPosition(streamers, Position.SUP),
-                extractByPosition(streamers, Position.COACH)
-        );
-
-        // 3. 최종 데이터 조립
-        DraftConfigResponse draftConfig = new DraftConfigResponse(pickOrder, coachResponses, streamersByLine);
-
-        return new DraftPlayStateResponse(room.getId(), room.getStatus(), draftConfig);
-    }
-
-    //    @Transactional
-//    public void configureAndStartRoom(Long roomId, String participantToken, RoomConfigureRequest request) {
-//        DraftRoom room = draftRoomRepository.findById(roomId)
-//                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 방입니다."));
-//
-//        DraftParticipant requestor = draftParticipantRepository.findByParticipantToken(participantToken)
-//                .orElseThrow(() -> new IllegalArgumentException("유효하지 않은 참여자입니다."));
-//
-//        if (!requestor.isHost()) {
-//            throw new IllegalArgumentException("방장만 드래프트를 시작할 수 있습니다.");
-//        }
-//
-//        if(!requestor.getRoomId().equals(roomId)) {
-//            throw new IllegalArgumentException("해당 방의 방장이 아닙니다.");
-//        }
-//
-//        room.updateSettings(request.teamCount(), request.teamSize());
-//
-//        List<DraftParticipant> participants = draftParticipantRepository.findAllByRoomId(roomId);
-//        if (participants.size() != room.getTeamCount()) {
-//            throw new IllegalStateException("설정된 팀 개수만큼 참여자가 모여야 시작할 수 있습니다.");
-//        }
-//
-//        room.start();
-//
-//        Collections.shuffle(participants);
-//        for (int i = 0; i < participants.size(); i++) {
-//            participants.get(i).assignTurnOrder(i);
-//        }
-//
-//        RoomStatusEvent event = RoomStatusEvent.builder()
-//                .code("SUCCESS")
-//                .roomStatus(room.getStatus())
-//                .redirectUrl("/drafts/" + roomId + "/play")
-//                .build();
-//
-//        applicationEventPublisher.publishEvent(
-//                DraftRoomStartedEvent.builder()
-//                        .roomId(roomId)
-//                        .payload(event)
-//                        .build()
-//        );
-//    }
 
 }
